app/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
celery[redis]==5.5.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
